move relationsToRepresentations to util file
