+++ src/app/+item-page/simple/item-types/shared/item.component.ts
@@ -1,5 +1,5 @@
 import { Component, Inject, OnInit } from '@angular/core';
-import { Observable ,  zip as observableZip, combineLatest as observableCombineLatest } from 'rxjs';
+import { combineLatest as observableCombineLatest, Observable, zip as observableZip } from 'rxjs';
 import { distinctUntilChanged, filter, flatMap, map } from 'rxjs/operators';
 import { ItemDataService } from '../../../../core/data/item-data.service';
 import { PaginatedList } from '../../../../core/data/paginated-list';
@@ -7,50 +7,10 @@ import { RemoteData } from '../../../../core/data/remote-data';
 import { RelationshipType } from '../../../../core/shared/item-relationships/relationship-type.model';
 import { Relationship } from '../../../../core/shared/item-relationships/relationship.model';
 import { Item } from '../../../../core/shared/item.model';
+import { MetadataRepresentation } from '../../../../core/shared/metadata-representation/metadata-representation.model';
 import { getRemoteDataPayload, getSucceededRemoteData } from '../../../../core/shared/operators';
 import { ITEM } from '../../../../shared/items/switcher/item-type-switcher.component';
-import { MetadataRepresentation } from '../../../../core/shared/metadata-representation/metadata-representation.model';
-import { ItemMetadataRepresentation } from '../../../../core/shared/metadata-representation/item/item-metadata-representation.model';
-import { MetadatumRepresentation } from '../../../../core/shared/metadata-representation/metadatum/metadatum-representation.model';
-import { of } from 'rxjs/internal/observable/of';
-import { MetadataValue } from '../../../../core/shared/metadata.models';
-import { compareArraysUsingIds } from './item-relationships-utils';
-
-/**
- * Operator for turning a list of relationships into a list of metadatarepresentations given the original metadata
- * @param thisId      The id of the parent item
- * @param itemType    The type of relation this list resembles (for creating representations)
- * @param metadata    The list of original Metadatum objects
- * @param ids         The ItemDataService to use for fetching Items from the Rest API
- */
-export const relationsToRepresentations = (thisId: string, itemType: string, metadata: MetadataValue[], ids: ItemDataService) =>
-  (source: Observable<Relationship[]>): Observable<MetadataRepresentation[]> =>
-    source.pipe(
-      flatMap((rels: Relationship[]) =>
-        observableZip(
-          ...metadata
-            .map((metadatum: any) => Object.assign(new MetadataValue(), metadatum))
-            .map((metadatum: MetadataValue) => {
-            if (metadatum.isVirtual) {
-              const matchingRels = rels.filter((rel: Relationship) => ('' + rel.id) === metadatum.virtualValue);
-              if (matchingRels.length > 0) {
-                const matchingRel = matchingRels[0];
-                let queryId = matchingRel.leftId;
-                if (matchingRel.leftId === thisId) {
-                  queryId = matchingRel.rightId;
-                }
-                return ids.findById(queryId).pipe(
-                  getSucceededRemoteData(),
-                  map((d: RemoteData<Item>) => Object.assign(new ItemMetadataRepresentation(), d.payload))
-                );
-              }
-            } else {
-              return of(Object.assign(new MetadatumRepresentation(itemType), metadatum));
-            }
-          })
-        )
-      )
-    );
+import { compareArraysUsingIds, relationsToRepresentations } from './item-relationships-utils';
 
 @Component({
   selector: 'ds-item',
